+++ jdaviz/configs/imviz/plugins/catalogs/catalogs.py
@@ -33,7 +33,15 @@ class Catalogs(PluginTemplateMixin, ViewerSelectMixin, HasFileImportSelect, Tabl
     * :meth:`~jdaviz.core.template_mixin.PluginTemplateMixin.show`
     * :meth:`~jdaviz.core.template_mixin.PluginTemplateMixin.open_in_tray`
     * :meth:`~jdaviz.core.template_mixin.PluginTemplateMixin.close_in_tray`
+    * :meth:`import_catalog`
     * :meth:`zoom_to_selected`
+    * :meth:`search`
+    * :attr:`max_sources`
+    * ``catalog`` (:class:`~jdaviz.core.template_mixin.SelectPluginComponent`)
+    * ``table`` (:class:`~jdaviz.core.template_mixin.Table`):
+      Table containing all search results.
+    * ``table_selected`` (:class:`~jdaviz.core.template_mixin.Table`):
+      Table containing all selected search results.
     """
     template_file = __file__, "catalogs.vue"
     uses_active_status = Bool(True).tag(sync=True)
@@ -62,7 +70,9 @@ class Catalogs(PluginTemplateMixin, ViewerSelectMixin, HasFileImportSelect, Tabl
     def user_api(self):
         return PluginUserApi(self, expose=('clear_table', 'export_table', 'import_catalog',
                                            'zoom_to_selected', 'select_rows',
-                                           'select_all', 'select_none'))
+                                           'select_all', 'select_none',
+                                           'catalog', 'max_sources', 'search',
+                                           'table', 'table_selected'))
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)